+++ tests/Datasets/UnlabeledTest.php
@@ -221,13 +221,26 @@ class UnlabeledTest extends TestCase
         $this->assertCount(3, $subset);
     }
 
-    public function test_merge_dataset()
+    public function test_prepend_dataset()
     {
         $this->assertCount(count($this->samples), $this->dataset);
 
         $dataset = new Unlabeled([['nice', 'furry', 'friendly']]);
 
-        $merged = $this->dataset->merge($dataset);
+        $merged = $this->dataset->prepend($dataset);
+
+        $this->assertCount(count($this->samples) + 1, $merged);
+
+        $this->assertEquals(['nice', 'furry', 'friendly'], $merged->row(0));
+    }
+
+    public function test_append_dataset()
+    {
+        $this->assertCount(count($this->samples), $this->dataset);
+
+        $dataset = new Unlabeled([['nice', 'furry', 'friendly']]);
+
+        $merged = $this->dataset->append($dataset);
 
         $this->assertCount(count($this->samples) + 1, $merged);