+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Unreleased
     - Pipeline and Persistent Model are now Probabilistic
     - Added stack method to dataset API
+    - Changed merge method on dataset to append and prepend
 
 - 0.0.8-alpha
     - Added Model Orchestra meta estimator